Reorganize some of `Gen`
The most noticeable change is the move of `BlockHash`, `TransactionId`
generators into `GenCoreProtocol`.

They were in `GenApiModel` but it was confusing as we use the types from
`org.alephium.protocol`

Some `*Entity` model were also moved from `Generators` into `GenDBModel`
