+++ app/src/test/scala/org/alephium/explorer/ExplorerSpec.scala
@@ -42,6 +42,8 @@ import org.alephium.api.{model, ApiError, ApiModelCodec}
 import org.alephium.api.{alphJsonBody => jsonBody}
 import org.alephium.explorer.ConfigDefaults._
 import org.alephium.explorer.GenApiModel._
+import org.alephium.explorer.GenCoreApi._
+import org.alephium.explorer.GenCoreProtocol._
 import org.alephium.explorer.Generators._
 import org.alephium.explorer.HttpFixture._
 import org.alephium.explorer.api._
