+++ app/src/test/scala/org/alephium/explorer/GenCoreProtocol.scala
@@ -18,18 +18,18 @@ package org.alephium.explorer
 
 import org.scalacheck.Gen
 
-import org.alephium.explorer.GenApiModel.hashGen
 import org.alephium.explorer.GenCommon._
 import org.alephium.protocol.{ALPH, Hash, PrivateKey, PublicKey, SignatureSchema}
-import org.alephium.protocol.config.GroupConfig
-import org.alephium.protocol.model.{ContractId, GroupIndex, NetworkId, ScriptHint}
+import org.alephium.protocol.model._
 import org.alephium.protocol.vm.LockupScript
 import org.alephium.util.AVector
 
 /** Generators for types supplied by Core `org.alephium.protocol` package */
 object GenCoreProtocol {
 
-  private implicit val groupConfig: GroupConfig = Generators.groupSetting.groupConfig
+  val hashGen: Gen[Hash]                     = Gen.const(()).map(_ => Hash.generate)
+  val blockHashGen: Gen[BlockHash]           = Gen.const(()).map(_ => BlockHash.generate)
+  val transactionHashGen: Gen[TransactionId] = hashGen.map(TransactionId.unsafe)
 
   val genNetworkId: Gen[NetworkId] =
     genBytePositive.map(NetworkId(_))
@@ -37,31 +37,36 @@ object GenCoreProtocol {
   def genNetworkId(exclude: NetworkId): Gen[NetworkId] =
     genNetworkId.filter(_ != exclude)
 
-  def keyPairGen(groupIndex: GroupIndex): (PrivateKey, PublicKey) = {
+  def keyPairGen(groupIndex: GroupIndex)(
+      implicit groupSetting: GroupSetting): (PrivateKey, PublicKey) = {
     val (privateKey, publicKey) = SignatureSchema.secureGeneratePriPub()
     val lockupScript            = LockupScript.p2pkh(Hash.hash(publicKey.bytes))
-    if (lockupScript.groupIndex == groupIndex) {
+    if (lockupScript.groupIndex(groupSetting.groupConfig) == groupIndex) {
       (privateKey, publicKey)
     } else {
       keyPairGen(groupIndex)
     }
   }
 
-  def publicKeyGen(groupIndex: GroupIndex): Gen[PublicKey] = keyPairGen(groupIndex)._2
+  def publicKeyGen(groupIndex: GroupIndex)(implicit groupSetting: GroupSetting): Gen[PublicKey] =
+    keyPairGen(groupIndex)._2
 
-  def p2pkhLockupGen(groupIndex: GroupIndex): Gen[LockupScript.P2PKH] =
+  def p2pkhLockupGen(groupIndex: GroupIndex)(
+      implicit groupSetting: GroupSetting): Gen[LockupScript.P2PKH] =
     for {
       publicKey <- publicKeyGen(groupIndex)
     } yield LockupScript.p2pkh(publicKey)
 
-  def p2mpkhLockupGen(groupIndex: GroupIndex): Gen[LockupScript.Asset] =
+  def p2mpkhLockupGen(groupIndex: GroupIndex)(
+      implicit groupSetting: GroupSetting): Gen[LockupScript.Asset] =
     for {
       numKeys   <- Gen.chooseNum(1, ALPH.MaxKeysInP2MPK)
       keys      <- Gen.listOfN(numKeys, publicKeyGen(groupIndex)).map(AVector.from)
       threshold <- Gen.choose(1, keys.length)
     } yield LockupScript.p2mpkh(keys, threshold).get
 
-  def p2mpkhLockupGen(n: Int, m: Int, groupIndex: GroupIndex): Gen[LockupScript.Asset] = {
+  def p2mpkhLockupGen(n: Int, m: Int, groupIndex: GroupIndex)(
+      implicit groupSetting: GroupSetting): Gen[LockupScript.Asset] = {
     assume(m <= n)
     for {
       publicKey0 <- publicKeyGen(groupIndex)
@@ -69,15 +74,17 @@ object GenCoreProtocol {
     } yield LockupScript.p2mpkh(publicKey0 +: moreKeys, m).get
   }
 
-  def p2shLockupGen(groupIndex: GroupIndex): Gen[LockupScript.Asset] = {
+  def p2shLockupGen(groupIndex: GroupIndex)(
+      implicit groupSetting: GroupSetting): Gen[LockupScript.Asset] = {
     hashGen
       .retryUntil { hash =>
-        ScriptHint.fromHash(hash).groupIndex.equals(groupIndex)
+        ScriptHint.fromHash(hash).groupIndex(groupSetting.groupConfig).equals(groupIndex)
       }
       .map(LockupScript.p2sh)
   }
 
-  def assetLockupGen(groupIndex: GroupIndex): Gen[LockupScript.Asset] = {
+  def assetLockupGen(groupIndex: GroupIndex)(
+      implicit groupSetting: GroupSetting): Gen[LockupScript.Asset] = {
     Gen.oneOf(
       p2pkhLockupGen(groupIndex),
       p2mpkhLockupGen(groupIndex),
@@ -85,17 +92,18 @@ object GenCoreProtocol {
     )
   }
 
-  def p2cLockupGen(groupIndex: GroupIndex): Gen[LockupScript.P2C] = {
+  def p2cLockupGen(groupIndex: GroupIndex)(
+      implicit groupSetting: GroupSetting): Gen[LockupScript.P2C] = {
     hashGen
       .retryUntil { hash =>
-        ScriptHint.fromHash(hash).groupIndex.equals(groupIndex)
+        ScriptHint.fromHash(hash).groupIndex(groupSetting.groupConfig).equals(groupIndex)
       }
       .map { hash =>
         LockupScript.p2c(ContractId.unsafe(hash))
       }
   }
 
-  def lockupGen(groupIndex: GroupIndex): Gen[LockupScript] = {
+  def lockupGen(groupIndex: GroupIndex)(implicit groupSetting: GroupSetting): Gen[LockupScript] = {
     Gen.oneOf(
       p2pkhLockupGen(groupIndex),
       p2mpkhLockupGen(groupIndex),
