+++ app/src/test/scala/org/alephium/explorer/persistence/queries/BlockQueriesSpec.scala
@@ -26,6 +26,7 @@ import slick.jdbc.PostgresProfile.api._
 import org.alephium.explorer.{AlephiumFutureSpec, GroupSetting}
 import org.alephium.explorer.ConfigDefaults._
 import org.alephium.explorer.GenApiModel._
+import org.alephium.explorer.GenCoreProtocol._
 import org.alephium.explorer.GenDBModel._
 import org.alephium.explorer.Generators._
 import org.alephium.explorer.api.model.{GroupIndex, Height}
