+++ app/src/test/scala/org/alephium/explorer/persistence/queries/InputUpdateQueriesSpec.scala
@@ -19,7 +19,7 @@ package org.alephium.explorer.persistence.queries
 import slick.jdbc.PostgresProfile.api._
 
 import org.alephium.explorer.AlephiumFutureSpec
-import org.alephium.explorer.Generators._
+import org.alephium.explorer.GenDBModel._
 import org.alephium.explorer.persistence.{DatabaseFixtureForEach, DBRunner}
 import org.alephium.explorer.persistence.schema.{InputSchema, OutputSchema}
 import org.alephium.explorer.persistence.schema.CustomJdbcTypes._
