+++ app/src/test/scala/org/alephium/explorer/service/TokenSupplyServiceSpec.scala
@@ -23,8 +23,8 @@ import scala.collection.immutable.ArraySeq
 import slick.jdbc.PostgresProfile.api._
 
 import org.alephium.explorer.{AlephiumFutureSpec, GroupSetting}
-import org.alephium.explorer.GenApiModel.transactionHashGen
-import org.alephium.explorer.Generators._
+import org.alephium.explorer.GenCoreProtocol.transactionHashGen
+import org.alephium.explorer.GenDBModel._
 import org.alephium.explorer.api.model._
 import org.alephium.explorer.cache.{BlockCache, TestBlockCache}
 import org.alephium.explorer.persistence._
